+++ transpiler/base.py
@@ -91,3 +91,17 @@ class LexerRule:
     def __init__(self, def_: Tag, regex: str) -> None:
         self.def_ = def_
         self.regex = regex
+
+
+class Token:
+    """
+    Minimal sensible unit of code sequence.
+    """
+
+    def __init__(self, def_: Tag, value: str, pos: int):
+        self.def_ = def_
+        self.value = value
+        self.pos = pos
+
+    def __str__(self):
+        return f'{self.value} {self.def_.value}'
