+++ transpiler/syntax_analyzer.py
@@ -1,0 +1,7 @@
+from typing import Iterable
+from transpiler.base import Tag, Token
+
+
+class SyntaxAnalyzer:
+    def __init__(self, tokens: Iterable[Token]):
+        self.tokens = tokens
